Core Prep to allow modular DataStores
Seperating the file storage classes from pywws.storage and allowing config to define a backend module to use.

Addition of clear(), update() and __iter__() methods to file storage classes to make data behave more like lists/dicts and to internalise the process of clearing entire datasets.

Adaptations to reprocess.py and other utility molules to account for these changes
